% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{get_mnpp.classtree}
\alias{get_mnpp.classtree}
\title{Get the MNPP for a Classification Tree}
\usage{
get_mnpp.classtree(z, data, Trt, Y, threshold)
}
\arguments{
\item{z}{a numeric vector of estimated CATEs from Step 1}

\item{data}{a data frame containing a response, binary treatment indicators,
and covariates.}

\item{Trt}{a string specifying the name of the column of \code{data}
contains the treatment indicators.}

\item{Y}{a string specifying the name of the column of \code{data}
contains the response.}

\item{threshold}{for "\code{step2 = 'classtree'}" only. The value against
which to test if the estimated individual treatment effect from Step 1 is
higher (TRUE) or lower (FALSE).}
}
\value{
the MNPP
}
\description{
Finds the lowest complexity parameter for a null regression tree fit
}
