% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{get_mnpp.ctree}
\alias{get_mnpp.ctree}
\title{Get the MNPP for a Conditional Inference Tree}
\usage{
get_mnpp.ctree(z, data, Trt, Y)
}
\arguments{
\item{z}{a numeric vector of estimated CATEs from Step 1}

\item{data}{a data frame containing a response, binary treatment indicators,
and covariates.}

\item{Trt}{a string specifying the name of the column of \code{data}
contains the treatment indicators.}

\item{Y}{a string specifying the name of the column of \code{data}
contains the response.}
}
\value{
the MNPP
}
\description{
Finds the lowest test statistic for a null conditional inference tree
}
