% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{get_mnpp.lasso}
\alias{get_mnpp.lasso}
\title{Get the MNPP for a Model fit via Lasso}
\usage{
get_mnpp.lasso(z, data, Trt, Y)
}
\arguments{
\item{z}{a numeric vector of estimated CATEs from Step 1}

\item{data}{a data frame containing a response, binary treatment indicators,
and covariates.}

\item{Trt}{a string specifying the name of the column of \code{data}
contains the treatment indicators.}

\item{Y}{a string specifying the name of the column of \code{data}
contains the response.}
}
\description{
Finds the lowest penalty parameter for a null lasso model.
}
