% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{get_mnpp.rtree}
\alias{get_mnpp.rtree}
\title{Get the MNPP for a Regression Tree}
\usage{
get_mnpp.rtree(z, data, Trt, Y)
}
\arguments{
\item{z}{a numeric vector of estimated CATEs from Step 1}

\item{data}{a data frame containing a response, binary treatment indicators,
and covariates.}

\item{Trt}{a string specifying the name of the column of \code{data}
contains the treatment indicators.}

\item{Y}{a string specifying the name of the column of \code{data}
contains the response.}
}
\value{
the MNPP
}
\description{
Finds the lowest complexity parameter for a null regression tree fit
}
