% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_1.R
\name{get_vt1}
\alias{get_vt1}
\title{Get the appropriate Step 1 estimation function associated with a method}
\usage{
get_vt1(step1)
}
\arguments{
\item{step1}{character strings specifying the Step 1 model. Supports
either "\code{lasso}", "\code{mars}", "\code{randomforest}", or
"\code{superlearner}".}
}
\value{
a function that estimates the CATE through Step 1 of Virtual Twins
}
\description{
Get the appropriate Step 1 estimation function associated with a method
}
