% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tehtuner_example}
\alias{tehtuner_example}
\title{Simulated example data}
\format{
A data frame with 200 rows and 12 columns:
\describe{
\item{Trt}{Binary treatment indicator}
\item{Y}{Continuous response}
\item{V1,V2,V3,V4,V5,V6,V7,V8}{Continuous covariates}
\item{V9,V10}{Binary covariates}
}
}
\usage{
tehtuner_example
}
\description{
Simulated data from a clinical trial with heterogenous treatment effects
where the CATE was a function of V1 and V9.
}
\keyword{datasets}
