% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{tunevt}
\alias{tunevt}
\title{Fit a tuned Virtual Twins model}
\usage{
tunevt(
  data,
  Y = "Y",
  Trt = "Trt",
  step1 = "randomforest",
  step2 = "rtree",
  alpha0,
  p_reps,
  keepz = FALSE,
  ...
)
}
\arguments{
\item{data}{a data frame containing a response, binary treatment indicators,
and covariates.}

\item{Y}{a string specifying the name of the column of \code{data}
contains the response.}

\item{Trt}{a string specifying the name of the column of \code{data}
contains the treatment indicators.}

\item{step1}{character strings specifying the Step 1 model. Supports
either "\code{lasso}", "\code{mars}", "\code{randomforest}", or
"\code{superlearner}".}

\item{step2}{a character string specifying the Step 2 model. Supports
"\code{lasso}", "\code{rtree}", or "\code{ctree}".}

\item{alpha0}{the nominal Type I error rate.}

\item{p_reps}{the number of permutations to run.}

\item{keepz}{logical. Should the estimated CATE from Step 1 be returned?}

\item{...}{additional arguments to the Step 1 model call.}
}
\value{
an object of class \code{"tunevt"}.

An object of class \code{"tunevt"} is a list containing at least the
following components:
\item{call}{the matched call}
\item{vtmod}{the model estimated by the given \code{step2} procedure fit
with the permuted tuning parameter for the estimated CATEs from the
\code{step1} model. See \code{\link{vt2_lasso}},
\code{\link{vt2_rtree}}, or \code{\link{vt2_ctree}} for specifics.}
\item{mnpp}{the MNPP for the estimated CATEs from Step 1.}
\item{theta_null}{a vector of the MNPPs from each permutation under
the null hypothesis.}
\item{z}{if \code{keepz = TRUE}, the estimated CATEs from the
\code{step1} model.}
}
\description{
\code{tunevt} fits a Virtual Twins model to estimate factors and subgroups
associated with differential treatment effects while controlling the Type I
error rate of falsely detecting at least one heterogeneous effect when the
treatment effect is uniform across the study population.
}
\details{
Virtual Twins is a two-step approach to detecting differential treatment
effects. Subjects' conditional average treatment effects (CATEs) are first
estimated in Step 1 using a flexible model. Then, a simple and interpretable
model is fit in Step 2 to model these estimated CATEs as a function of the
covariates.

The Step 2 model is dependent on some tuning parameter. This parameter is
selected to control the Type I error rate by permuting the data under the
null hypothesis of a constant treatment effect and identifying the minimal
null penalty parameter (MNPP), which is the smallest penalty parameter that
yields a Step 2 model with no covariate effects. The \code{1-alpha0} quantile
of the distribution of is then used to fit the Step 2 model on the original
data.
}
\examples{
data(tehtuner_example)
tunevt(
  tehtuner_example, step1 = "lasso", step2 = "rtree",
  alpha0 = 0.05, p_reps = 10, keepz = TRUE
)

}
\references{
{

\insertRef{foster_subgroup_2011}{tehtuner}

}
}
