% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_input.R
\name{validate_alpha0}
\alias{validate_alpha0}
\title{Check if alpha0 is a valid input to tunevt}
\usage{
validate_alpha0(data, alpha0)
}
\arguments{
\item{data}{a data frame containing a response, binary treatment indicators,
and covariates.}

\item{alpha0}{the nominal Type I error rate.}
}
\value{
TRUE if \code{alpha0} is a valid input. Errors otherwise.
}
\description{
Check if alpha0 is a valid input to tunevt
}
