% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_input.R
\name{validate_p_reps}
\alias{validate_p_reps}
\title{Check if p_reps is a valid input to tunevt}
\usage{
validate_p_reps(data, p_reps)
}
\arguments{
\item{data}{a data frame containing a response, binary treatment indicators,
and covariates.}

\item{p_reps}{the number of permutations to run.}
}
\value{
TRUE if \code{p_reps} is a valid input. Errors otherwise.
}
\description{
Check if p_reps is a valid input to tunevt
}
