% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_input.R
\name{validate_Trt}
\alias{validate_Trt}
\title{Check if Trt is a valid input to tunevt}
\usage{
validate_Trt(data, Trt)
}
\arguments{
\item{data}{a data frame containing a response, binary treatment indicators,
and covariates.}

\item{Trt}{a string specifying the name of the column of \code{data}
contains the treatment indicators.}
}
\value{
TRUE if \code{Trt} is a valid input. Errors otherwise.
}
\description{
Check if Trt is a valid input to tunevt
}
