% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_input.R
\name{validate_Y}
\alias{validate_Y}
\title{Check if Y is a valid input to tunevt}
\usage{
validate_Y(data, Y)
}
\arguments{
\item{data}{a data frame containing a response, binary treatment indicators,
and covariates.}

\item{Y}{a string specifying the name of the column of \code{data}
contains the response.}
}
\value{
TRUE if \code{Y} is a valid input. Errors otherwise.
}
\description{
Check if Y is a valid input to tunevt
}
