% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_1.R
\name{vt1_lasso}
\alias{vt1_lasso}
\title{Estimate the CATE Using the Lasso for Step 1 of Virtual Twins}
\usage{
vt1_lasso(data, Trt, Y, ...)
}
\arguments{
\item{data}{a data frame containing a response, binary treatment indicators,
and covariates.}

\item{Trt}{a string specifying the name of the column of \code{data}
contains the treatment indicators.}

\item{Y}{a string specifying the name of the column of \code{data}
contains the response.}

\item{...}{additional arguments to \code{cv.glmnet}}
}
\value{
Estimated CATEs for each subject in \code{data}.
}
\description{
Estimate the CATE Using the Lasso for Step 1 of Virtual Twins
}
\seealso{
Other VT Step 1 functions: 
\code{\link{vt1_mars}()},
\code{\link{vt1_rf}()},
\code{\link{vt1_super}()}
}
\concept{VT Step 1 functions}
