% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute.R
\name{permute}
\alias{permute}
\title{Generate a dataset with permuated treatment indicators}
\usage{
permute(data, Trt, Y, zbar)
}
\arguments{
\item{data}{a data frame containing a response, binary treatment indicators,
and covariates.}

\item{Trt}{a string specifying the name of the column of \code{data}
contains the treatment indicators.}

\item{Y}{a string specifying the name of the column of \code{data}
contains the response.}

\item{zbar}{the estimated marginal treatment effect}
}
\value{
a permuted dataset of the same size as \code{data}
}
\description{
Sets the marginal treatment effect to zero and then permute all treatment
indicators.
}
